Added abstract class item and made sub-abstract classes for weapon, artifact, and perk (and an enum for buff). Will be able to create instances such as shortsword, regeneration, etc.

+++ core/src/com/raginggoose/roguetrails/item/Weapon.java
@@ -3,7 +3,7 @@ package com.raginggoose.roguetrails.item;
 public abstract class Weapon extends Item {
     public abstract int getDamage();
     public abstract void setDamage(int damage);
-    public abstract int getCooldown();
+    public abstract double getCooldown();
     public abstract void setCooldown(double cooldown);
     public abstract int getRange();
     public abstract void setRange(int range);
